+++ static/app/views/replays/detail/breadcrumbs/index.tsx
@@ -8,7 +8,7 @@ import {useReplayContext} from 'sentry/components/replays/replayContext';
 import useJumpButtons from 'sentry/components/replays/useJumpButtons';
 import {t} from 'sentry/locale';
 import useCrumbHandlers from 'sentry/utils/replays/hooks/useCrumbHandlers';
-import useExtractedDomNodes from 'sentry/utils/replays/hooks/useExtractedDomNodes';
+import useExtractDomNodes from 'sentry/utils/replays/hooks/useExtractDomNodes';
 import useVirtualizedInspector from 'sentry/views/replays/detail//useVirtualizedInspector';
 import BreadcrumbFilters from 'sentry/views/replays/detail/breadcrumbs/breadcrumbFilters';
 import BreadcrumbRow from 'sentry/views/replays/detail/breadcrumbs/breadcrumbRow';
@@ -30,8 +30,10 @@ const cellMeasurer = {
 function Breadcrumbs() {
   const {currentTime, replay} = useReplayContext();
   const {onClickTimestamp} = useCrumbHandlers();
-  const {data: frameToExtraction, isFetching: isFetchingExtractions} =
-    useExtractedDomNodes({replay});
+
+  const {data: frameToExtraction, isFetching: isFetchingExtractions} = useExtractDomNodes(
+    {replay}
+  );
 
   const startTimestampMs = replay?.getStartTimestampMs() ?? 0;
   const frames = replay?.getChapterFrames();
