+++ static/app/views/replays/detail/memoryPanel/domNodesChart.tsx
@@ -16,7 +16,7 @@ import {getFormattedDate} from 'sentry/utils/dates';
 import {axisLabelFormatter} from 'sentry/utils/discover/charts';
 import domId from 'sentry/utils/domId';
 import formatReplayDuration from 'sentry/utils/duration/formatReplayDuration';
-import type {DomNodeChartDatapoint} from 'sentry/utils/replays/countDomNodes';
+import type {DomNodeChartDatapoint} from 'sentry/utils/replays/hooks/useCountDomNodes';
 
 interface Props
   extends Pick<ReturnType<typeof useReplayContext>, 'currentTime' | 'setCurrentTime'> {
